fix: only load MDX on the server
